refactor(kernel): extract Arrow cell/JSON rendering to untagged arrowscan (N5/N6)
Move ScanCell (scalar + nested→JSON grammar: native float32, exact decimals,
time.Time formatting, list/map/struct) out of the cgo-tagged kernel package into
a new pure-Go internal/arrowscan package. rows.go delegates via
arrowscan.ScanCell. The renderer imports no C, so this is a pure move.

Why: the rendering rules are the parity-critical contract both backends must
agree on, but every kernel-package test is //go:build cgo && databricks_kernel
and dead in CI. Relocating the tests to arrowscan makes them run in the default
CGO_ENABLED=0 matrix (N6) — TestScanCellScalars/Nested/TimestampLocation now
guard the native-float32 (N3/M2) and exact-decimal rendering on every build. The
grammar is single-sourced for the kernel side (N5).

Kernel-specific tests (error mapping, bad-connection, bound-params rejection)
stay in the kernel package. Verified: default suite + arrowscan race pass; kernel
unit tests pass; live Thrift-parity + all data types still byte-identical.

Note: the Thrift arrowbased path still has its own nested renderer (built on its
columnValues container abstraction, not raw arrow.Array). Having it delegate to
arrowscan too — the remaining half of N5 — is a separate, higher-risk refactor of
the primary production path; deferred.
